requirements.txt
python-telegram-bot==13.13
environs==9.5.0
google-cloud-dialogflow==2.15.0
vk-api==11.9.9